requirements.txt
# E-Clinic Authentication Service Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OAuth & Google Integration
authlib==1.3.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client (for MSG91)
requests==2.31.0
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time utilities
python-dateutil==2.8.2
